+++ cypress/support/create-user.ts
@@ -7,8 +7,8 @@
 import { installGlobals } from "@remix-run/node";
 import { parse } from "cookie";
 
-import { createUser } from "~/models/user.server";
-import { createUserSession } from "~/session.server";
+import { createUserSession } from "~/utils/session.server";
+import { createUser } from "~/utils/user.server";
 
 installGlobals();
