+++ cypress/support/delete-user.ts
@@ -6,7 +6,7 @@
 import { PrismaClientKnownRequestError } from "@prisma/client/runtime";
 import { installGlobals } from "@remix-run/node";
 
-import { prisma } from "~/db.server";
+import { prisma } from "~/utils/db.server";
 
 installGlobals();
